refactor: VIdeo 목록 추출하여 클래스 파일로 생성, 하드코딩 대체
