+++ app/src/main/java/com/kotlinisgood/boomerang/ui/home/VideoGallery.kt
@@ -1,0 +1,64 @@
+package com.kotlinisgood.boomerang.ui.home
+
+import android.content.ContentResolver
+import android.content.ContentUris
+import android.net.Uri
+import android.provider.MediaStore
+import android.util.Log
+import java.util.concurrent.TimeUnit
+
+class VideoGallery(private val contentResolver: ContentResolver) {
+
+    fun loadVideos(): MutableList<ExternalVideoDTO> {
+        val videoList = mutableListOf<ExternalVideoDTO>()
+
+        val projection = arrayOf(
+            MediaStore.Video.Media._ID,
+            MediaStore.Video.Media.DISPLAY_NAME,
+            MediaStore.Video.Media.DURATION,
+            MediaStore.Video.Media.SIZE
+        )
+
+        val selection = "${MediaStore.Video.Media.DURATION} <= ?"
+        val selectionArgs = arrayOf(
+            TimeUnit.MILLISECONDS.convert(1, TimeUnit.MINUTES).toString()
+        )
+        val sortOrder = "${MediaStore.Video.Media.DISPLAY_NAME} ASC"
+
+        val query = contentResolver.query(
+            MediaStore.Video.Media.EXTERNAL_CONTENT_URI,
+            projection,
+            selection,
+            selectionArgs,
+            sortOrder
+        )
+
+        query?.use { cursor ->
+            val idColumn = cursor.getColumnIndexOrThrow(MediaStore.Video.Media._ID)
+            val nameColumn =
+                cursor.getColumnIndexOrThrow(MediaStore.Video.Media.DISPLAY_NAME)
+            val durationColumn =
+                cursor.getColumnIndexOrThrow(MediaStore.Video.Media.DURATION)
+            val sizeColumn = cursor.getColumnIndexOrThrow(MediaStore.Video.Media.SIZE)
+
+            while (cursor.moveToNext()) {
+                val id = cursor.getLong(idColumn)
+                val name = cursor.getString(nameColumn)
+                val duration = cursor.getInt(durationColumn)
+                val size = cursor.getInt(sizeColumn)
+
+                val contentUri: Uri = ContentUris.withAppendedId(
+                    MediaStore.Video.Media.EXTERNAL_CONTENT_URI,
+                    id
+                )
+
+                videoList += ExternalVideoDTO(contentUri, name, duration, size)
+            }
+        }
+
+        Log.i(HomeFragment.TAG, videoList.size.toString())
+        videoList.forEach { Log.i(HomeFragment.TAG, "$it") }
+        return videoList
+    }
+
+}
